+++ src/btree/backfill.cc
@@ -13,22 +13,6 @@
 values from the leaf nodes. */
 static const int MAX_CONCURRENT_VALUE_LOADS = 16;
 
-RDB_IMPL_SERIALIZABLE_1_FOR_CLUSTER(backfill_pre_item_t, range);
-RDB_IMPL_SERIALIZABLE_3_FOR_CLUSTER(backfill_item_t::pair_t, key, recency, value);
-RDB_IMPL_SERIALIZABLE_3_FOR_CLUSTER(backfill_item_t,
-    range, pairs, min_deletion_timestamp);
-
-void backfill_item_t::mask_in_place(const key_range_t &m) {
-    range = range.intersection(m);
-    std::vector<pair_t> new_pairs;
-    for (auto &&pair : pairs) {
-        if (m.contains_key(pair.key)) {
-            new_pairs.push_back(std::move(pair));
-        }
-    }
-    pairs = std::move(new_pairs);
-}
-
 /* `convert_to_right_bound()` is suitable for converting `btree_key_t *`s in the format
 of `right_incl` or `left_excl_or_null` into a `key_range_t::right_bound_t`. */
 key_range_t::right_bound_t convert_to_right_bound(const btree_key_t *rightmost_before) {
